feat(telemetry): rename useModeTelemetry->useDiscoveryTelemetry, add runtimes
