+++ web/src/App.tsx
@@ -7,7 +7,7 @@ import { CliDrawer } from './components/CliDrawer'
 import { getTheme, type Theme } from './lib/prefs'
 import { safeGet } from './lib/safeStorage'
 import { trackAction, trackView } from './lib/telemetry'
-import { useModeTelemetry } from './hooks/useModeTelemetry'
+import { useDiscoveryTelemetry } from './hooks/useDiscoveryTelemetry'
 
 const SIDEBAR_COLLAPSED_KEY = 'devdash.sidebarCollapsed'
 
@@ -41,7 +41,7 @@ export function App() {
   // Mounted here (not per-route) so the ['apps'] react-query stays active on
   // every route, keeping the `modes` telemetry context live app-wide — do
   // not "optimize" this by moving it closer to where apps are displayed.
-  useModeTelemetry()
+  useDiscoveryTelemetry()
 
   useEffect(() => {
     trackAction('app_startup')
